+++ miuix/src/commonMain/kotlin/top/yukonga/miuix/kmp/color/space/Hsv.kt
@@ -8,15 +8,15 @@ import androidx.compose.ui.graphics.Color.Companion.hsv
 
 /**
  * HSV representation with normalized, user-friendly ranges.
- * - `h`: hue in degrees [0, 360)
- * - `s`: saturation in percent [0.0, 100.0]
- * - `v`: value/brightness in percent [0.0, 100.0]
+ * - `h`: hue in degrees [0, 360]
+ * - `s`: saturation in percent [0, 100]
+ * - `v`: value/brightness in percent [0, 100]
  */
-data class Hsv(val h: Double, val s: Double, val v: Double) {
+data class Hsv(val h: Float, val s: Float, val v: Float) {
     fun toColor(alpha: Float = 1f): Color {
-        val hue = (((h % 360.0) + 360.0) % 360.0).toFloat()
-        val sN = (s / 100.0).coerceIn(0.0, 1.0).toFloat()
-        val vN = (v / 100.0).coerceIn(0.0, 1.0).toFloat()
+        val hue = (((h % 360f) + 360f) % 360f)
+        val sN = (s / 100f).coerceIn(0f, 1f)
+        val vN = (v / 100f).coerceIn(0f, 1f)
         return hsv(hue, sN, vN, alpha)
     }
 }
